refactor(examples): split notification lab
