+++ examples/notifications/main.ts
@@ -1,725 +1,14 @@
 import { pathToFileURL } from 'node:url';
-import { initDefaultContext } from '@flyingrobots/bijou-node';
-import { box, kbd, resolveClock, stringToSurface } from '@flyingrobots/bijou';
-import {
-  activateFocusedNotification,
-  createFramedApp,
-  createKeyMap,
-  createNotificationState,
-  cycleNotificationFocus,
-  dismissFocusedNotification,
-  dismissNotification,
-  isMouseMsg,
-  notificationsNeedTick,
-  pushNotification,
-  relocateNotifications,
-  quit,
-  renderNotificationStack,
-  run,
-  tick,
-  tickNotifications,
-  trimNotificationsToViewport,
-  stripAnsi,
-  type Cmd,
-  type FramePage,
-  type MouseMsg,
-  type NotificationHistoryFilter,
-  type NotificationPlacement,
-  type NotificationSpec,
-  type NotificationState,
-  type NotificationTone,
-  type NotificationVariant,
-} from '../../packages/bijou-tui/src/index.js';
+import { run } from '../../packages/bijou-tui/src/index.js';
+import { app } from './notification-demo-entry.js';
 
-export const ctx = initDefaultContext();
+export { createNotificationDemoApp } from './notification-demo-create.js';
+export { app } from './notification-demo-entry.js';
+export { ctx } from './notification-demo-options.js';
 
-const NOTIFICATION_TICK_MS = 40;
-
-const VARIANTS: readonly NotificationVariant[] = ['ACTIONABLE', 'INLINE', 'TOAST'];
-const TONES: readonly NotificationTone[] = ['INFO', 'SUCCESS', 'WARNING', 'ERROR'];
-const PLACEMENTS: readonly NotificationPlacement[] = [
-  'UPPER_LEFT',
-  'UPPER_RIGHT',
-  'LOWER_LEFT',
-  'LOWER_RIGHT',
-  'TOP_CENTER',
-  'BOTTOM_CENTER',
-  'CENTER',
-];
-
-const HISTORY_FILTERS: readonly NotificationHistoryFilter[] = [
-  'ALL',
-  'ACTIONABLE',
-  'ERROR',
-  'WARNING',
-  'SUCCESS',
-  'INFO',
-];
-
-const DURATION_OPTIONS = [
-  { label: 'default', value: undefined },
-  { label: 'persistent', value: null },
-  { label: '2.5s', value: 2_500 },
-  { label: '5.0s', value: 5_000 },
-  { label: '9.0s', value: 9_000 },
-] as const;
-
-const COMPACT_DEMO_WIDTH = 52;
-const COMPACT_DEMO_HEIGHT = 16;
-const COMPACT_OVERLAY_MARGIN = 1;
-const DEFAULT_OVERLAY_MARGIN = 2;
-const DEMO_NOTIFICATION_GAP = 1;
-
-function isCompactDemoViewport(width: number, height: number): boolean {
-  return width < COMPACT_DEMO_WIDTH || height < COMPACT_DEMO_HEIGHT;
-}
-
-function demoOverlayMargin(width: number, height: number): number {
-  return isCompactDemoViewport(width, height)
-    ? COMPACT_OVERLAY_MARGIN
-    : DEFAULT_OVERLAY_MARGIN;
-}
-
-type Msg =
-  | MouseMsg
-  | { type: 'spawn-notification' }
-  | { type: 'cycle-variant' }
-  | { type: 'cycle-tone' }
-  | { type: 'cycle-placement' }
-  | { type: 'cycle-duration' }
-  | { type: 'toggle-action' }
-  | { type: 'toggle-wrap' }
-  | { type: 'set-history-filter'; filter: NotificationHistoryFilter }
-  | { type: 'focus-next' }
-  | { type: 'focus-prev' }
-  | { type: 'activate-focused' }
-  | { type: 'dismiss-notification' }
-  | { type: 'notification-tick' }
-  | { type: 'key-observed'; key: string; route: string }
-  | { type: 'quit-app' }
-  | { type: 'notification-action'; ordinal: number };
-
-const pageKeyMap = createKeyMap<Msg>()
-  .bind('n', 'Spawn notification', { type: 'spawn-notification' })
-  .bind('v', 'Cycle variant', { type: 'cycle-variant' })
-  .bind('t', 'Cycle tone', { type: 'cycle-tone' })
-  .bind('l', 'Cycle placement', { type: 'cycle-placement' })
-  .bind('d', 'Cycle duration', { type: 'cycle-duration' })
-  .bind('a', 'Toggle action button', { type: 'toggle-action' })
-  .bind('w', 'Toggle text wrap', { type: 'toggle-wrap' })
-  .bind('j', 'Focus next actionable notification', { type: 'focus-next' })
-  .bind('k', 'Focus previous actionable notification', { type: 'focus-prev' })
-  .bind('enter', 'Run focused notification action', { type: 'activate-focused' })
-  .bind('x', 'Dismiss focused/latest notification', { type: 'dismiss-notification' });
-
-interface PageModel {
-  readonly notifications: NotificationState<Msg>;
-  readonly notificationLoopActive: boolean;
-  readonly variantIndex: number;
-  readonly toneIndex: number;
-  readonly placementIndex: number;
-  readonly durationIndex: number;
-  readonly actionEnabled: boolean;
-  readonly wrapText: boolean;
-  readonly historyFilterIndex: number;
-  readonly nextOrdinal: number;
-  readonly lastHandledInput: string;
-  readonly log: readonly string[];
-}
-
-function createInitialPageModel(): PageModel {
-  return {
-    notifications: createNotificationState<Msg>(),
-    notificationLoopActive: false,
-    variantIndex: 0,
-    toneIndex: 0,
-    placementIndex: 3,
-    durationIndex: 0,
-    actionEnabled: true,
-    wrapText: true,
-    historyFilterIndex: 0,
-    nextOrdinal: 1,
-    lastHandledInput: 'none',
-    log: [
-      'Notification lab ready.',
-      'Press n to spawn a new notification.',
-    ],
-  };
-}
-
-function currentVariant(model: PageModel) { return at(VARIANTS, model.variantIndex); }
-function currentTone(model: PageModel) { return at(TONES, model.toneIndex); }
-function currentPlacement(model: PageModel) { return at(PLACEMENTS, model.placementIndex); }
-function currentDuration(model: PageModel) { return at(DURATION_OPTIONS, model.durationIndex); }
-function currentHistoryFilter(model: PageModel) { return at(HISTORY_FILTERS, model.historyFilterIndex); }
-function n(value: number) { return String(value); }
-function at<T>(v: readonly T[], i: number) { const value = v[i]; if (value === undefined) throw new Error('Bad notification state.'); return value; }
-function mouse(msg: MouseMsg) { return `mouse:${n(msg.col)},${n(msg.row)}`; }
-
-function toneCopy(tone: NotificationTone): { readonly title: string; readonly message: string } {
-  switch (tone) {
-    case 'INFO':
-      return {
-        title: 'Background sync ready',
-        message: 'Fresh data is available for review.',
-      };
-    case 'SUCCESS':
-      return {
-        title: 'Release shipped cleanly',
-        message: 'All checks passed and artifacts are live.',
-      };
-    case 'WARNING':
-      return {
-        title: 'Queue pressure rising',
-        message: 'Latency is trending upward on the worker pool.',
-      };
-    case 'ERROR':
-      return {
-        title: 'Deploy blocked',
-        message: 'The runtime failed to boot the latest candidate.',
-      };
-  }
-}
-
-function actionLabelForTone(tone: NotificationTone): string {
-  switch (tone) {
-    case 'INFO':
-      return 'Open details';
-    case 'SUCCESS':
-      return 'Share result';
-    case 'WARNING':
-      return 'Inspect queue';
-    case 'ERROR':
-      return 'Retry deploy';
-  }
-}
-
-function formatDurationLabel(value: number | null | undefined): string {
-  if (value === undefined) return 'default';
-  if (value === null) return 'persistent';
-  return `${(value / 1000).toFixed(1)}s`;
-}
-
-function appendLog(model: PageModel, message: string): PageModel {
-  return {
-    ...model,
-    log: [message, ...model.log].slice(0, 12),
-  };
-}
-
-function recordInput(model: PageModel, key: string, message: string): PageModel {
-  return appendLog({
-    ...model,
-    lastHandledInput: key,
-  }, `[${key}] ${message}`);
-}
-
-function renderControlsPane(model: PageModel, width: number, notificationCtx = ctx): string {
-  const focused = model.notifications.focusedId == null ? 'none' : `#${n(model.notifications.focusedId)}`;
-  const lines = [
-    'Notification Lab',
-    '',
-    `Variant : ${currentVariant(model)}`,
-    `Tone    : ${currentTone(model)}`,
-    `Next at : ${currentPlacement(model)}`,
-    `Cycle   : ${at(PLACEMENTS, (model.placementIndex + 1) % PLACEMENTS.length)}`,
-    `Stay    : ${currentDuration(model).label}`,
-    `Action  : ${model.actionEnabled ? 'enabled' : 'disabled'}`,
-    `Wrap    : ${model.wrapText ? 'enabled' : 'disabled'}`,
-    `Stack   : ${n(model.notifications.items.length)}`,
-    `History : ${n(model.notifications.history.length)}`,
-    `Review  : shift+n (${currentHistoryFilter(model)})`,
-    `Focus   : ${focused}`,
-    `Last key: ${model.lastHandledInput}`,
-    '',
-    `${kbd('n')} spawn`,
-    `${kbd('v')} cycle variant`,
-    `${kbd('t')} cycle tone`,
-    `${kbd('l')} cycle placement`,
-    `${kbd('d')} cycle duration`,
-    `${kbd('a')} toggle action`,
-    `${kbd('w')} toggle wrap`,
-    `${kbd('shift+n')} shell notification center`,
-    `${kbd('j')} / ${kbd('k')} focus action`,
-    `${kbd('enter')} run action`,
-    `${kbd('x')} dismiss focused/latest`,
-    `${kbd('/')} search • ${kbd('f2')} settings`,
-    '',
-    'Try stacking a few in the same corner, then',
-    'flip placements to compare the anchor behavior.',
-  ];
-
-  return box(lines.join('\n'), {
-    width,
-    title: 'Controls',
-    overflow: model.wrapText ? 'wrap' : 'truncate',
-    ctx: notificationCtx,
-  });
-}
-
-function renderLogPane(model: PageModel, width: number, notificationCtx = ctx): string {
-  const lines = [
-    'Recent events',
-    '',
-    ...model.log,
-  ];
-
-  return box(lines.join('\n'), {
-    width,
-    title: 'Activity',
-    overflow: model.wrapText ? 'wrap' : 'truncate',
-    ctx: notificationCtx,
-  });
-}
-
-function applyNotificationState(
-  model: PageModel,
-  notifications: NotificationState<Msg>,
-  notificationCtx = ctx,
-  commands: readonly Cmd<Msg>[] = [],
-  forceTick = false,
-): [PageModel, Cmd<Msg>[]] {
-  const clock = resolveClock(notificationCtx);
-  const trimmed = trimNotificationsToViewport(notifications, {
-    screenWidth: notificationCtx.runtime.columns,
-    screenHeight: Math.max(0, notificationCtx.runtime.rows - 2),
-    margin: demoOverlayMargin(
-      notificationCtx.runtime.columns,
-      Math.max(0, notificationCtx.runtime.rows - 2),
-    ),
-    gap: DEMO_NOTIFICATION_GAP,
-    ctx: notificationCtx,
-  }, clock.now());
-  const needsTick = notificationsNeedTick(trimmed);
-  const nextModel: PageModel = {
-    ...model,
-    notifications: trimmed,
-    notificationLoopActive: needsTick,
-  };
-  const shouldScheduleTick = needsTick && (forceTick || !model.notificationLoopActive);
-
-  return [
-    nextModel,
-    shouldScheduleTick
-      ? [...commands, tick(NOTIFICATION_TICK_MS, { type: 'notification-tick' })]
-      : [...commands],
-  ];
-}
-
-function spawnConfiguredNotification(
-  model: PageModel,
-  notificationCtx = ctx,
-): [PageModel, Cmd<Msg>[]] {
-  const clock = resolveClock(notificationCtx);
-  const ordinal = model.nextOrdinal;
-  const variant = currentVariant(model);
-  const tone = currentTone(model);
-  const placement = currentPlacement(model);
-  const duration = currentDuration(model);
-  const copy = toneCopy(tone);
-  const spec: NotificationSpec<Msg> = {
-    title: `${copy.title} #${n(ordinal)}`,
-    message: `${copy.message} ${variant} @ ${placement} • ${duration.label}`,
-    variant,
-    tone,
-    placement,
-    durationMs: duration.value,
-    action: variant === 'ACTIONABLE' && model.actionEnabled
-      ? {
-        label: actionLabelForTone(tone),
-        payload: { type: 'notification-action', ordinal },
-      }
-      : undefined,
-    overflow: model.wrapText ? 'wrap' : 'truncate',
-  };
-
-  const notifications = pushNotification(model.notifications, spec, clock.now());
-  const nextModel = appendLog({
-    ...model,
-    lastHandledInput: 'n',
-    notifications,
-    nextOrdinal: ordinal + 1,
-  }, `[n] Spawned ${variant} #${n(ordinal)} at ${placement} (${formatDurationLabel(duration.value)}).`);
-
-  return applyNotificationState(nextModel, notifications, notificationCtx);
-}
-
-function dismissCurrentNotification(
-  model: PageModel,
-  notificationCtx = ctx,
-): [PageModel, Cmd<Msg>[]] {
-  const nowMs = resolveClock(notificationCtx).now();
-  const latest = model.notifications.items.at(-1);
-  const notifications = model.notifications.focusedId != null
-    ? dismissFocusedNotification(model.notifications, nowMs)
-    : (latest == null ? model.notifications : dismissNotification(model.notifications, latest.id, nowMs));
-
-  if (notifications === model.notifications) return [model, []];
-
-  const nextModel = recordInput({
-    ...model,
-    notifications,
-  }, 'x', 'Dismissed a notification.');
-
-  return applyNotificationState(nextModel, notifications, notificationCtx);
-}
-
-function notificationRegion(notificationCtx = ctx) {
-  return {
-    row: 2,
-    col: 0,
-    width: notificationCtx.runtime.columns,
-    height: Math.max(0, notificationCtx.runtime.rows - 2),
-  };
-}
-
-function resolveDemoNotificationMouseTarget(
-  model: PageModel,
-  msg: MouseMsg,
-  notificationCtx = ctx,
+if (
+  process.argv[1] != null &&
+  import.meta.url === pathToFileURL(process.argv[1]).href
 ) {
-  const overlays = renderNotificationStack(model.notifications, {
-    screenWidth: notificationCtx.runtime.columns,
-    screenHeight: notificationCtx.runtime.rows,
-    region: notificationRegion(notificationCtx),
-    margin: 2,
-    gap: 1,
-    ctx: notificationCtx,
-  });
-
-  for (const overlay of [...overlays].reverse()) {
-    const surface = overlay.surface;
-    if (
-      surface == null
-      || msg.col < overlay.col
-      || msg.col >= overlay.col + surface.width
-      || msg.row < overlay.row
-      || msg.row >= overlay.row + surface.height
-    ) {
-      continue;
-    }
-
-    const item = model.notifications.items.find((candidate) =>
-      stripAnsi(overlay.content).includes(candidate.title)
-    );
-    if (item == null) continue;
-
-    const localCol = msg.col - overlay.col;
-    const localRow = msg.row - overlay.row;
-    if (localRow === 0 && localCol === surface.width - 2) {
-      return { item, kind: 'dismiss' as const };
-    }
-    if (item.variant === 'ACTIONABLE' && localRow === surface.height - 1 && localCol >= 2) {
-      return { item, kind: 'action' as const };
-    }
-    return { item, kind: 'body' as const };
-  }
-
-  return undefined;
-}
-
-function handleNotificationMouse(
-  model: PageModel,
-  msg: MouseMsg,
-  notificationCtx = ctx,
-): [PageModel, Cmd<Msg>[]] {
-  if (msg.action !== 'press' || msg.button !== 'left') return [model, []];
-
-  const target = resolveDemoNotificationMouseTarget(model, msg, notificationCtx);
-
-  if (target == null) return [model, []];
-
-  if (target.kind === 'dismiss') {
-    const nowMs = resolveClock(notificationCtx).now();
-    const notifications = dismissNotification(model.notifications, target.item.id, nowMs);
-    const nextModel = appendLog({
-      ...model,
-      notifications,
-      lastHandledInput: mouse(msg),
-    }, `[mouse] Dismissed notification #${n(target.item.id)}.`);
-    return applyNotificationState(nextModel, notifications, notificationCtx);
-  }
-
-  if (target.kind === 'action') {
-    const nowMs = resolveClock(notificationCtx).now();
-    const notifications = dismissNotification(model.notifications, target.item.id, nowMs);
-    const action = target.item.action;
-    const nextModel = appendLog({
-      ...model,
-      notifications,
-      lastHandledInput: mouse(msg),
-    }, action == null
-      ? `[mouse] Dismissed notification #${n(target.item.id)}.`
-      : `[mouse] Activated notification #${n(target.item.id)}.`);
-
-    if (action == null) {
-      return applyNotificationState(nextModel, notifications, notificationCtx);
-    }
-
-    const actionCmd: Cmd<Msg> = () => action.payload;
-    return applyNotificationState(nextModel, notifications, notificationCtx, [actionCmd]);
-  }
-
-  return [model, []];
-}
-
-function seedDemoNotifications(model: PageModel, notificationCtx = ctx): PageModel {
-  const entries = [
-    {
-      title: 'Deploy blocked',
-      message: 'Retryable actionable notice in the upper right.',
-      variant: 'ACTIONABLE' as const,
-      tone: 'ERROR' as const,
-      placement: 'UPPER_RIGHT' as const,
-      durationMs: null,
-      actionLabel: 'Retry deploy',
-    },
-    {
-      title: 'Queue pressure rising',
-      message: 'Inline notice centered at the top edge.',
-      variant: 'INLINE' as const,
-      tone: 'WARNING' as const,
-      placement: 'TOP_CENTER' as const,
-      durationMs: 5_000,
-    },
-    {
-      title: 'Release shipped cleanly',
-      message: 'Toast variant stacked near the lower-right anchor.',
-      variant: 'TOAST' as const,
-      tone: 'SUCCESS' as const,
-      placement: 'LOWER_RIGHT' as const,
-      durationMs: 4_000,
-    },
-  ] as const;
-
-  let next = model;
-  let nowMs = resolveClock(notificationCtx).now();
-
-  for (const entry of entries) {
-    const ordinal = next.nextOrdinal;
-    const notifications = pushNotification(next.notifications, {
-      title: `${entry.title} #${n(ordinal)}`,
-      message: entry.message,
-      variant: entry.variant,
-      tone: entry.tone,
-      placement: entry.placement,
-      durationMs: entry.durationMs,
-      action: 'actionLabel' in entry
-        ? {
-          label: entry.actionLabel,
-          payload: { type: 'notification-action', ordinal },
-        }
-        : undefined,
-      overflow: next.wrapText ? 'wrap' : 'truncate',
-    }, nowMs);
-
-    next = appendLog({
-      ...next,
-      notifications,
-      nextOrdinal: ordinal + 1,
-    }, `Seeded ${entry.variant} #${n(ordinal)} for automated smoke rendering.`);
-
-    nowMs += 60;
-  }
-
-  return next;
-}
-
-export function createNotificationDemoApp(
-  notificationCtx = ctx,
-  options: { readonly autoDemo?: boolean } = {},
-) {
-  const autoDemo = options.autoDemo ?? (
-    notificationCtx.runtime.env('CI') === '1' || process.env.CI === '1'
-  );
-
-  const page: FramePage<PageModel, Msg> = {
-    id: 'notifications',
-    title: 'Notifications',
-    init() {
-      const seeded = autoDemo
-        ? seedDemoNotifications(createInitialPageModel(), notificationCtx)
-        : createInitialPageModel();
-      const [model, cmds] = applyNotificationState(seeded, seeded.notifications, notificationCtx);
-      return [
-        model,
-        autoDemo ? [...cmds, tick(1_600, { type: 'quit-app' })] : cmds,
-      ];
-    },
-    update(msg, model) {
-      if (isMouseMsg(msg)) {
-        return handleNotificationMouse(model, msg, notificationCtx);
-      }
-
-      switch (msg.type) {
-      case 'spawn-notification':
-        return spawnConfiguredNotification(model, notificationCtx);
-      case 'cycle-variant':
-        return [recordInput({
-          ...model,
-          variantIndex: (model.variantIndex + 1) % VARIANTS.length,
-        }, 'v', `Variant -> ${at(VARIANTS, (model.variantIndex + 1) % VARIANTS.length)}.`), []];
-      case 'cycle-tone':
-        return [recordInput({
-          ...model,
-          toneIndex: (model.toneIndex + 1) % TONES.length,
-        }, 't', `Tone -> ${at(TONES, (model.toneIndex + 1) % TONES.length)}.`), []];
-      case 'cycle-placement':
-      {
-        const nextPlacementIndex = (model.placementIndex + 1) % PLACEMENTS.length;
-        const nextPlacement = at(PLACEMENTS, nextPlacementIndex);
-        const notifications = relocateNotifications(
-          model.notifications,
-          nextPlacement,
-          resolveClock(notificationCtx).now(),
-        );
-        const nextModel = appendLog({
-          ...model,
-          lastHandledInput: 'l',
-          placementIndex: nextPlacementIndex,
-          notifications,
-        }, `[l] Placement -> ${nextPlacement}; active notifications relocated.`);
-        return applyNotificationState(nextModel, notifications, notificationCtx);
-      }
-      case 'cycle-duration':
-        return [recordInput({
-          ...model,
-          durationIndex: (model.durationIndex + 1) % DURATION_OPTIONS.length,
-        }, 'd', `Duration -> ${at(DURATION_OPTIONS, (model.durationIndex + 1) % DURATION_OPTIONS.length).label}.`), []];
-      case 'toggle-action':
-        return [recordInput({
-          ...model,
-          actionEnabled: !model.actionEnabled,
-        }, 'a', `Action button ${!model.actionEnabled ? 'enabled' : 'disabled'}.`), []];
-      case 'toggle-wrap':
-        return [recordInput({
-          ...model,
-          wrapText: !model.wrapText,
-        }, 'w', `Wrap ${!model.wrapText ? 'enabled' : 'disabled'}.`), []];
-      case 'set-history-filter':
-        if (msg.filter === currentHistoryFilter(model)) return [model, []];
-        return [recordInput({
-          ...model,
-          historyFilterIndex: HISTORY_FILTERS.indexOf(msg.filter),
-        }, 'f', `History filter -> ${msg.filter}.`), []];
-      case 'focus-next':
-        return [recordInput({
-          ...model,
-          notifications: cycleNotificationFocus(model.notifications, 1),
-        }, 'j', 'Focused next actionable notification.'), []];
-      case 'focus-prev':
-        return [recordInput({
-          ...model,
-          notifications: cycleNotificationFocus(model.notifications, -1),
-        }, 'k', 'Focused previous actionable notification.'), []];
-      case 'dismiss-notification':
-        return dismissCurrentNotification(model, notificationCtx);
-      case 'activate-focused': {
-        const result = activateFocusedNotification(model.notifications, resolveClock(notificationCtx).now());
-        let nextModel = {
-          ...model,
-          notifications: result.state,
-        };
-
-        if (result.payload?.type === 'notification-action') {
-          nextModel = recordInput(nextModel, 'enter', `Action fired from notification #${n(result.payload.ordinal)}.`);
-        }
-
-        return applyNotificationState(nextModel, nextModel.notifications, notificationCtx);
-      }
-      case 'notification-action':
-        return [appendLog(model, `Notification #${n(msg.ordinal)} delivered its action payload.`), []];
-      case 'notification-tick': {
-        const notifications = tickNotifications(model.notifications, resolveClock(notificationCtx).now());
-        return applyNotificationState({
-          ...model,
-          notifications,
-        }, notifications, notificationCtx, [], true);
-      }
-      case 'key-observed':
-        return [appendLog({
-          ...model,
-          lastHandledInput: `${msg.key}:${msg.route}`,
-        }, `[key ${msg.key}] route=${msg.route}`), []];
-      case 'quit-app':
-        return [model, [quit()]];
-      case 'pulse':
-        return [model, []];
-      default:
-        return [model, []];
-    }
-    },
-    keyMap: pageKeyMap,
-    layout(model) {
-      return {
-        kind: 'grid',
-        gridId: 'notification-lab',
-        columns: [38, '1fr'],
-        rows: ['1fr'],
-        areas: ['controls activity'],
-        gap: 1,
-        cells: {
-          controls: {
-            kind: 'pane',
-            paneId: 'controls',
-            render: (width, height) => stringToSurface(renderControlsPane(model, width, notificationCtx), width, height),
-          },
-          activity: {
-            kind: 'pane',
-            paneId: 'activity',
-            render: (width, height) => stringToSurface(renderLogPane(model, width, notificationCtx), width, height),
-          },
-        },
-      };
-    },
-  };
-
-  return createFramedApp<PageModel, Msg>({
-    title: 'Bijou Notification Lab',
-    pages: [page],
-    initialColumns: notificationCtx.runtime.columns,
-    initialRows: notificationCtx.runtime.rows,
-    keyPriority: 'page-first',
-    observeKey: (msg, route) => ({
-      type: 'key-observed',
-      key: `${msg.ctrl ? 'ctrl+' : ''}${msg.alt ? 'alt+' : ''}${msg.shift ? 'shift+' : ''}${msg.key}`,
-      route,
-    }),
-    enableCommandPalette: true,
-    notificationCenter: ({ pageModel }) => ({
-      title: 'Notification center',
-      state: pageModel.notifications,
-      filters: HISTORY_FILTERS,
-      activeFilter: currentHistoryFilter(pageModel),
-      onFilterChange: (filter) => ({ type: 'set-history-filter', filter }),
-    }),
-    overlayFactory(frame) {
-      const paneRects = [...frame.paneRects.values()];
-      const region = paneRects.length === 0
-        ? frame.screenRect
-        : {
-          col: Math.min(...paneRects.map((rect) => rect.col)),
-          row: Math.min(...paneRects.map((rect) => rect.row)),
-          width: Math.max(...paneRects.map((rect) => rect.col + rect.width)) - Math.min(...paneRects.map((rect) => rect.col)),
-          height: Math.max(...paneRects.map((rect) => rect.row + rect.height)) - Math.min(...paneRects.map((rect) => rect.row)),
-        };
-
-      const overlays = [...renderNotificationStack(frame.pageModel.notifications, {
-        screenWidth: frame.screenRect.width,
-        screenHeight: frame.screenRect.height,
-        region,
-        margin: 2,
-        gap: 1,
-        ctx: notificationCtx,
-      })];
-
-      return overlays;
-    },
-  });
-}
-
-export const app = createNotificationDemoApp();
-
-if (process.argv[1] != null && import.meta.url === pathToFileURL(process.argv[1]).href) {
   void run(app, { mouse: true });
 }
