+++ examples/notifications/notification-demo-entry.ts
@@ -1,0 +1,3 @@
+import { createNotificationDemoApp } from './notification-demo-create.js';
+
+export const app = createNotificationDemoApp();
